feat: added category validator

+++ src/main/java/tiimae/webshop/iprwc/controller/CategoryController.java
@@ -8,6 +8,7 @@ import tiimae.webshop.iprwc.constants.ApiConstant;
 import tiimae.webshop.iprwc.mapper.CategoryMapper;
 import tiimae.webshop.iprwc.models.Category;
 import tiimae.webshop.iprwc.service.ApiResponseService;
+import tiimae.webshop.iprwc.validators.CategoryValidator;
 
 import java.util.UUID;
 
@@ -16,10 +17,12 @@ public class CategoryController {
 
     private CategoryDAO categoryDAO;
     private CategoryMapper categoryMapper;
+    private CategoryValidator categoryValidator;
 
-    public CategoryController(CategoryDAO categoryDAO, CategoryMapper categoryMapper) {
+    public CategoryController(CategoryDAO categoryDAO, CategoryMapper categoryMapper, CategoryValidator categoryValidator) {
         this.categoryDAO = categoryDAO;
         this.categoryMapper = categoryMapper;
+        this.categoryValidator = categoryValidator;
     }
 
     @GetMapping(ApiConstant.getOneCategories)
@@ -37,6 +40,12 @@ public class CategoryController {
     @PostMapping(ApiConstant.getAllCategories)
     @ResponseBody
     public ApiResponseService post(@RequestBody CategoryDTO categoryDTO) {
+        final String validate = this.categoryValidator.validateDTO(categoryDTO);
+
+        if (validate != null) {
+            return new ApiResponseService(HttpStatus.BAD_REQUEST, validate);
+        }
+
         final Category category = this.categoryMapper.toCategory(categoryDTO);
 
         return new ApiResponseService(HttpStatus.CREATED, this.categoryDAO.create(category));
@@ -46,6 +55,18 @@ public class CategoryController {
     @ResponseBody
     @CrossOrigin
     public ApiResponseService put(@PathVariable UUID categoryId, @RequestBody CategoryDTO categoryDTO) {
+        final String validate = this.categoryValidator.validateDTO(categoryDTO);
+
+        if (validate != null) {
+            return new ApiResponseService(HttpStatus.BAD_REQUEST, validate);
+        }
+
+        final String validateId = this.categoryValidator.validateId(categoryId);
+
+        if (validateId != null) {
+            return new ApiResponseService(HttpStatus.BAD_REQUEST, validateId);
+        }
+
         return new ApiResponseService(HttpStatus.ACCEPTED, this.categoryDAO.update(categoryId, categoryDTO));
     }
 
@@ -53,6 +74,12 @@ public class CategoryController {
     @ResponseBody
     @CrossOrigin
     public ApiResponseService delete(@PathVariable UUID categoryId) {
+        final String validateId = this.categoryValidator.validateId(categoryId);
+
+        if (validateId != null) {
+            return new ApiResponseService(HttpStatus.BAD_REQUEST, validateId);
+        }
+
         this.categoryDAO.delete(categoryId);
 
         return new ApiResponseService(HttpStatus.ACCEPTED, "Category has been removed");
